user/requirements.txt
flask
python_dotenv
supabase
flask_cors
